+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Allow clickable views inside list items
+- Increased background task timeout to 30s
+
 ## [0.10.0] - 2023-12-20
 
 ### Added